moved pagebuilder to documents as we probably will not have  other builders
